♻️ refactor: refactor code, create subcomponents, organize folder structure, and use CSS modules
