+++ .vscode/settings.json
@@ -3,4 +3,5 @@
   "editor.codeActionsOnSave": {
     "source.fixAll.eslint": true
   },
+  "scss.lint.unknownAtRules": "ignore"
 }
