+++ README.md
@@ -1,13 +1,10 @@
 # Memory Game
 
 Memory Game is a memorization game created for a challenge. This project is built with Next.js 13, uses hooks and context to persist state, and is set up with ESLint (Airbnb style), lint-staged, Prettier, and Husky.
-<p align="center">
 
 ![Next JS](https://img.shields.io/badge/Next-black?style=for-the-badge&logo=next.js&logoColor=white)        ![SASS](https://img.shields.io/badge/SASS-hotpink.svg?style=for-the-badge&logo=SASS&logoColor=white) ![TailwindCSS](https://img.shields.io/badge/tailwindcss-%2338B2AC.svg?style=for-the-badge&logo=tailwind-css&logoColor=white) ![Docker](https://img.shields.io/badge/docker-%230db7ed.svg?style=for-the-badge&logo=docker&logoColor=white) ![Netlify](https://img.shields.io/badge/netlify-%23000000.svg?style=for-the-badge&logo=netlify&logoColor=#00C7B7)
-</p>
-<p align="center">
+
 [![Live Demo](https://img.shields.io/badge/demo-online-green.svg)](https://eclectic-kelpie-73a18e.netlify.app) [![Commitizen friendly](https://img.shields.io/badge/commitizen-friendly-brightgreen.svg)](http://commitizen.github.io/cz-cli/) [![Node.js Version](https://img.shields.io/badge/node-%3E%3D16.8.0-brightgreen)](https://nodejs.org/en/download/) [![License: GPL v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
-</p>
 
 ## Features
 
@@ -21,7 +18,7 @@ Memory Game is a memorization game created for a challenge. This project is buil
 
 You can find a live version of the project at this link: [https://eclectic-kelpie-73a18e.netlify.app](https://eclectic-kelpie-73a18e.netlify.app).
 
-# Requirements
+## Requirements
 
 - Node.js 16.8.0 or newer
 - Docker Compose (Optional)
