--- src/components/game/Card.module.scss
@@ -1,51 +1,0 @@
-.card {
-  &__container {
-    width: 100%;
-    max-width: 7rem;
-    height: 9rem;
-    perspective: 1000px;
-    margin: 0 auto;
-    cursor: pointer;
-    transform-style: preserve-3d;
-    transition: all 0.3s ease-in-out;
-    color: white;
-    font-size: 48px;
-    font-weight: 600;
-
-    &:hover {
-      transform: scale(1.1);
-    }
-  }
-
-  &__inner {
-    width: 100%;
-    height: 100%;
-    position: relative;
-    transform-style: preserve-3d;
-    transition: all 0.5s ease-in-out;
-  }
-
-  &__side {
-    width: 100%;
-    height: 100%;
-    position: absolute;
-    backface-visibility: hidden;
-    border-radius: 5px;
-    box-shadow: 0 4px 6px rgba(50, 50, 93, 0.11), 0 1px 3px rgba(0, 0, 0, 0.08);
-  }
-
-  &__front {
-    background-color: #fff;
-  }
-
-  &__back {
-    background-color: #151f38;
-    transform: rotateY(180deg);
-  }
-
-  &--is-flipped {
-    .card__inner {
-      transform: rotateY(180deg);
-    }
-  }
-}
