+++ src/pages/_app.jsx
@@ -1,36 +1,14 @@
-import Head from 'next/head';
-
-import Image from 'next/image';
-// eslint-disable-next-line camelcase
-import { DM_Sans } from 'next/font/google';
 import { PlayerProvider } from '../contexts/PlayerContext';
+import Layout from '../components/Layout';
 
 import '@/styles/globals.scss';
 
-const dm = DM_Sans({ subsets: ['latin'], weight: ['400', '500', '700'] });
-
 export default function App({ Component, pageProps }) {
   return (
-    <>
-      <Head>
-        <title>Memory Game</title>
-        <meta name="description" content="Juega al juego de memoria con tus amigos" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <meta charSet="utf-8" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-
-      <PlayerProvider>
-        <div className={`${dm.className} subpixel-antialiased`}>
-          <Component {...pageProps} />
-          <footer className="flex flex-col items-center justify-center flex-1 py-4 mt-4">
-            <Image src="/images/modyo.svg" width={100} height={100} alt="Modyo" />
-            <p className="text-xs text-slate-300 hover:text-slate-500 ease-in-out duration-300 mt-2">
-              Crafted by <span className="font-bold">Fernando Figueroa</span>
-            </p>
-          </footer>
-        </div>
-      </PlayerProvider>
-    </>
+    <PlayerProvider>
+      <Layout>
+        <Component {...pageProps} />
+      </Layout>
+    </PlayerProvider>
   );
 }
